cleanup ag-charts-enterprise main.ts
